updates predict next purchase
