+++ docs/source/examples/demo/predict_next_purchase/__init__.py
@@ -1,13 +1,12 @@
 import os
-import tarfile
-
 import pandas as pd
 import requests
-
+import tarfile
+from demo import utils
 from tqdm import tqdm
 
+URL = r'https://s3.amazonaws.com/instacart-datasets/instacart_online_grocery_shopping_2017_05_01.tar.gz'
 PWD = os.path.dirname(__file__)
-PATH = os.path.join(PWD, 'instacart_2017_05_01')
 
 
 def add_time(df, start='2015-01-01'):
@@ -16,14 +15,11 @@ def add_time(df, start='2015-01-01'):
 
     def process_orders(user):
         orders = user.groupby('order_id')
-
         days = orders.days_since_prior_order.first()
         days = days.cumsum().apply(timedelta, string='{}d')
         days = days.add(pd.Timestamp(start))
-
         hour_of_day = orders.order_hour_of_day.first()
         hour_of_day = hour_of_day.apply(timedelta, string='{}h')
-
         order_time = days.add(hour_of_day)
         return order_time
 
@@ -43,50 +39,26 @@ def add_time(df, start='2015-01-01'):
     return df
 
 
-def load_orders(path=None, nrows=1000000):
-    if not exists(): download()
-
-    path = path or PATH
+def load_data(nrows=1000000):
+    output = os.path.join(PWD, 'download')
+    path = os.path.join(output, 'instacart_2017_05_01')
+    if not os.path.exists(path): utils.download(URL, output)
 
     file = os.path.join(path, 'order_products__prior.csv')
     order_products = pd.read_csv(file, nrows=nrows)
-
-    file = os.path.join(path, 'orders.csv')
-    orders = pd.read_csv(file, nrows=nrows)
-
-    file = os.path.join(path, 'departments.csv')
-    departments = pd.read_csv(file)
-
     file = os.path.join(path, 'products.csv')
     products = pd.read_csv(file)
+    file = os.path.join(path, 'departments.csv')
+    departments = pd.read_csv(file)
+    file = os.path.join(path, 'orders.csv')
+    orders = pd.read_csv(file, nrows=nrows)
 
     df = order_products.merge(products).merge(departments).merge(orders)
     df = df.pipe(add_time)
     return df
 
 
-def exists():
-    return os.path.exists(PATH)
-
-
-def download():
-    url = r'https://s3.amazonaws.com/instacart-datasets/instacart_online_grocery_shopping_2017_05_01.tar.gz'
-    response = requests.get(url, stream=True)
-    assert response.status_code == 200, "unable to download data"
-
-    bar_format = "Downloaded: {n}MB / {total}MB -{rate_fmt}, "
-    bar_format += "Elapsed: {elapsed}, Remaining: {remaining}, Progress: {l_bar}{bar}"
-    total = round(int(response.headers.get('content-length', 0)) / 1e+6)
-    data = response.iter_content(chunk_size=int(1e+6))
-    data = tqdm(data, total=total, unit="MB", bar_format=bar_format)
-
-    tar = os.path.join(PWD, 'data.tar.gz')
-    with open(tar, 'wb') as file:
-        for chunk in data:
-            file.write(chunk)
-
-    with tarfile.open(tar, "r:gz") as file:
-        file.extractall('data')
-
-    os.remove(tar)
-    response.close()
+def load_sample():
+    path = os.path.join(PWD, 'sample.csv')
+    df = pd.read_csv(path, parse_dates=['order_time'], index_col='id')
+    return df
